Move to Ubuntu 12.10 to get newer OVS packages and libvirt.
Can now use stock packages for libvirt, and I'm hoping the OVS
packages will be more stable since I seemed to have random
issues with the 1.4.0 ones.

Works, but still unclear if the right kernel module from the
openvswitch-datapath-dkms is activated on install. Need one
more try from scratch on that but leaving hosts up for stability
testing on 12.10 first.